Test mvp run without GLCM
