--- .github/workflows/comment.yml
@@ -1,16 +1,0 @@
-name: Model Training
-
-on:
-  push:
-    branches: [ "main" ]
-  pull_request:
-
-jobs:
-  build:
-    runs-on: self-hosted
-    container: docker://ghcr.io/iterative/cml:0-dvc2-base1-gpu
-    steps:
-      - uses: actions/checkout@v3
-      - name: Try comment
-        run: |
-          cml comment update --target=pr README.md
