+++ .github/workflows/model.yml
@@ -1,11 +1,7 @@
 name: Model Training
 
 on:
-  workflow_dispatch:
-
-#  push:
-#    branches: [ "main" ]
-#  pull_request:
+  pull_request:
 
 jobs:
   build:
@@ -38,14 +34,17 @@ jobs:
       - name: 'Set up Cloud SDK'
         uses: 'google-github-actions/setup-gcloud@v1'
 
-      #     We don't necessarily need to install the CUDA version of torch, so we'll do with cpu
-#      - name: Install dependencies through poetry
-#        run: |
-#          python -m pip install --upgrade pip
-#          poetry install --with dev,glcm
-#          pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
+      - name: 'Set up WandB'
+        run: |
+          echo "export WANDB_API_KEY=${{ secrets.WANDB_API_KEY }}" >> $GITHUB_ENV
 
       - name: Test with pytest through poetry
         run: |
-          pytest 
+          pytest tests/model_tests/chestnut_dec_may
 
+      - name: Comment results
+        run: |
+          cml comment update \
+          --target=pr \
+          --token ${{ secrets.GITHUB_TOKEN }} \
+          tests/model_tests/chestnut_dec_may/report.md
