+++ tests/model_tests/chestnut_dec_may/preprocess.py
@@ -1,10 +1,12 @@
 import numpy as np
 import torch
-from glcm_cupy import Features
+
+# from glcm_cupy import Features
 from torchvision.transforms.v2 import Resize
 
 from frdc.models import FaceNet
-from frdc.preprocess.glcm_padded import append_glcm_padded_cached
+
+# from frdc.preprocess.glcm_padded import append_glcm_padded_cached
 from frdc.preprocess.scale import scale_normal_per_band, scale_0_1_per_band
 
 
@@ -24,16 +26,17 @@ def segment_preprocess(ar: np.ndarray) -> torch.Tensor:
 
     # Add a small epsilon to avoid upper bound of 1.0
     ar = scale_0_1_per_band(ar, epsilon=0.001)
-    ar = append_glcm_padded_cached(
-        ar,
-        step_size=7,
-        bin_from=1,
-        bin_to=128,
-        radius=3,
-        features=(Features.MEAN,),
-    )
-    # We can then scale normal for better neural network convergence
+    # ar = append_glcm_padded_cached(
+    #     ar,
+    #     step_size=7,
+    #     bin_from=1,
+    #     bin_to=128,
+    #     radius=3,
+    #     features=(Features.MEAN,),
+    # )
+    # # We can then scale normal for better neural network convergence
     ar = scale_normal_per_band(ar)
+    ar = np.rollaxis(ar, axis=2)
 
     # TODO: Doesn't seem like we have any channel preprocessing here.
     # ar = np.stack([
